refactor(service): rename workflows + remove read-write mode from pipelineWorkflow
- Rename pipeline.ts → pipeline-workflow.ts (same export name pipelineWorkflow)
- Rename pipeline-google-sheets.ts → google-sheet-pipeline-workflow.ts,
  export pipelineGoogleSheetsWorkflow → googleSheetPipelineWorkflow
- Remove read-write mode (separate readLoop/writeLoop) from pipelineWorkflow;
  the workflow now only handles the syncImmediate (Postgres) path
- Remove readIntoQueue/writeFromQueue proxy exports from workflows/_shared.ts
- Update all call sites (app.ts, app.test.ts, workflow.test.ts)

Committed-By-Agent: claude
