+++ apps/service/src/temporal/workflows/_shared.ts
@@ -41,7 +41,7 @@ export const { setup, teardown } = proxyActivities<SyncActivities>({
   retry: retryPolicy,
 })
 
-export const { syncImmediate, readIntoQueue, writeFromQueue } = proxyActivities<SyncActivities>({
+export const { syncImmediate } = proxyActivities<SyncActivities>({
   startToCloseTimeout: '10m',
   heartbeatTimeout: '2m',
   retry: retryPolicy,
